refactor(main): extract the REPL phases and the last subcommands (3,840 → 1,267)
Second and final mechanical pass. What is left in main.rs is `main`'s clap dispatch, the two
REPL loops, and a handful of small shared helpers — no subsystem remains to lift out.

  src/repl/postturn.rs    572  auto-compact · secretary · memory learning · persona evolution
  src/cli/run_cmds.rs     457  prompt-size · crawl · reach · models · chat · agent · workflow
  src/repl/turn.rs        399  one shared agent turn, from registry to finish_turn
  src/repl/background.rs  289  health chip · reconcile sweep · aside worker
  src/repl/startup.rs     241  identity banner · onboarding · MCP trust · id migration
  src/cli/coop_cmd.rs     222  `aizen zone/team/work`
  src/cli/memory_cmd.rs   222  `aizen memory …`
  src/repl/input_pre.rs   154  `#remember` · `!shell` · `@file` expansion
  src/ui/provider_ui.rs   124  `/model` · provider picker · profile switching
  src/core/endpoint.rs     80  endpoint resolution + the shared HTTP client

These are not leaves the way the first pass's nine were — they reach back into the crate root
and into each other — so unlike `features/slash_handlers.rs` and `ui/menus.rs`, which paper over
that with `use crate::*`, every one of them imports what it uses by name. That was the point of
splitting: 43 explicit imports say what these modules actually depend on, where a glob would have
hidden it. Only the items genuinely reached from outside their module are `pub(crate)` — a first
blanket pass published 69, and 12 came back off once measured.

Two things the compiler cannot see, so they are recorded here:

- a crate-root item that is merely private is visible to the WHOLE crate (every module descends
  from the root). That is why the pre-move code compiled with almost no `pub` anywhere, and why
  moving a function into a submodule is the step that first forces its visibility to be stated.
- `src/tests/main_suite.rs` reaches the root through `use super::*`, so it inherited main.rs's
  imports. Two of those (`anyhow!`, `core::config`) stopped existing there once the subcommands
  left; the test module now imports them itself instead of main.rs carrying them for it.

Verified: 1635 passed / 0 failed / 2 ignored (unchanged); cargo fmt --all -- --check clean;
clippy 125 warnings, identical to the baseline — the two that now point into src/repl and src/cli
are pre-existing lints that moved with their code.
