+++ src/cli/mod.rs
@@ -3,7 +3,10 @@
 
 pub mod agents_cmd;
 pub mod apps;
+pub mod coop_cmd;
+pub mod memory_cmd;
 pub mod persona_cmd;
+pub mod run_cmds;
 pub mod sessions;
 pub mod skill_cmd;
 pub mod time;
