+++ src/ui/mod.rs
@@ -15,6 +15,7 @@ pub mod menus;
 pub mod mermaid;
 pub mod moonscape;
 pub mod plain_input;
+pub mod provider_ui;
 pub mod spinner;
 pub mod splash;
 pub mod theme;
